convert to typescript Fix bug in uploading mutual groups info.
